Bug 1806770 - Use ImapUid as return type for nsImapFlagAndUidState.GetHighestNonDeletedUID(). r=mkmelin
Differential Revision: https://phabricator.services.mozilla.com/D295422

--HG--
extra : rebase_source : 9d7cff2d64d133cc311c909b9275fc0634392d04

+++ mailnews/imap/src/nsImapProtocol.cpp
@@ -4610,7 +4610,7 @@ void nsImapProtocol::PeriodicBiff() {
       AppendUid(fetchStr, highestUid);
       fetchStr.AppendLiteral(":*");
       FetchMessage(fetchStr, kFlags);
-      if (((uint32_t)m_flagState->GetHighestNonDeletedUID() >= highestUid) &&
+      if ((m_flagState->GetHighestNonDeletedUID() >= highestUid) &&
           m_flagState->IsLastMessageUnseen())
         m_currentBiffState = nsIMsgFolder::nsMsgBiffState_NewMail;
       else
